configure: Install view.
(better subject, maybe)

I still think "view" should be available to users, so the only thing between the user and "view" is they have to compile Wine theirselves and that should change.
It's very hard to find good wmf/emf viewers for Linux(and most likely other OS), but "view" is a great program for people who need to take a look at wmf/emf files.