wined3d: Avoid clearing the current wined3d context in WineD3D_CreateFakeGLContext().
There should be no need to clear the thread's current wined3d context as long
as we properly restore the corresponding GL stuff. This avoids hitting the
context restore code in context_enter() when we have e.g. two d3d9 devices,
one of which is inactive, because now we can see that the GL context that's
active belongs to another wined3d context, and we don't need to bother
saving and restoring it.