quartz: Remove the seek position table from the MPEG splitter.
It takes a good 0.5 seconds for MPEGSplitter_pre_connect to iterate over the
file and build the table for a standard 2 or 3 minute MP3. This causes
unsightly hiccups when an app tries to play a new file, particularly in games,
as the pre-connect needs to happen synchronously to tell whether it can
succeed or not.