user32: Fix mouseRect computation for starting position when sizing window.
The mouseRect is the limit of the mouse position to keep the window between
its minimum and maximum sizes.  The computation was correct if the mouse
started precisely on the window border.  But if the mouse is not exactly at
the border, then the mouse limit has to account for that, because the window
will be resized based on the distance from the starting point.

(cherry picked from commit c7d2bcfb6114fb13eeee8d0513f4587472d20019)