opengl32: Allow multiple extensions to support the same function.
This allows that gl functions are available for different extensions,
e.g. "glVertexP2ui" which is new GL_ARB_vertex_type_2_10_10_10_rev and
also in GL_VERSION_3_3. It is valid to get that function for each gl
version when GL_ARB_vertex_type_2_10_10_10_rev is in the extension list.
On the other hand that function is available, when the gl version is >= 3.3.