d3d9/tests: Extend the FPU setup test.
Games created by GameMaker modify the FPU control word to enable
division by zero exceptions. The game Risk of Rain, which is written
with GameMaker, furthermore sets fogstart == fogend, which triggers an
exception when the GL driver attempts to calculate 1 / (fogend -
fogstart). This test adds some evidence that we are not supposed to
temporarily restore the FPU CW if D3DCREATE_FPU_PRESERVE is used.