shell32: Use QueryInterface in FolderView_GetFolder.
test_IFolderView has a test that shows that the shell folder's reference
count before and after the call to IFolderView_GetFolder is the same.
However this is misleading because it is not a property of the
IFolderView_GetFolder method, but of the desktop shell folder object
that's used in the folder view. Windows <=2003 uses a cached object and
its AddRef/Release methods return constants.