msidb: Add stub tool for manipulating MSI databases.
The "Windows SDK Components for Windows Installer Developers" has a
command line tool called msidb that is incredibly useful for creating,
editing, and exporting MSI installer databases, think of it as
winemsibuilder on steroids.  This patch series implements much of the
functionality of the msidb tool, maintains compatible CLI flags, and
the underlying MSI functionality necessary to support these features.
Jacek expressed an interest in having these patches resurrected for
use by the Gecko build scripts and Austin's VS builds of Valgrind.
With this patch series all the existing winemsibuilder functionality
is available, plus the ability to drop streams, export the
_SummaryInformation table, and export binary streams (Binary/Icon
tables).  A big feature of the implementation is that it allows you to
edit existing installer databases, rather than just creating new ones.

Signed-off-by: Erich E. Hoover <erich.e.hoover@gmail.com>
Signed-off-by: Hans Leidekker <hans@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>