loader: Add the CFBundleAllowMixedLocalizations key to the embedded Info.plist to allow Cocoa to localize its strings.
The loader itself is not a bundled application, so it has no localizations as
far as the Mac APIs are concerned.  Its CFBundleDevelopmentRegion is English,
so that's the preferred localization that the Mac APIs choose for it.

Cocoa manages the Enter Full Screen menu item that the Mac driver creates.  By
default, Cocoa prefers not to mix localizations.  Therefore, even though Cocoa
contains localized strings for "Enter Full Screen", it was just using the
English one for Wine, regardless of system language.

Setting this key tells it to use the non-English localizations from Cocoa even
though the loader's own preferred localization is English.

Signed-off-by: Ken Thomases <ken@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>